[Attack Discovery] Add a Generations Control Center to the Attacks page (#277472)
## Summary

This PR introduces a **Generations Control Center** to the Attacks page:
a single flyout that lists running and recent attack discovery
generations, and lets the user drill into the workflow execution details
of any generation — all without stacking flyouts.

A new **Generations** button is added to the Attacks page header (before
the **Run** and **Schedule** actions). Clicking it opens one flyout that
follows a master–detail pattern:

- **List view** — shows the latest generations (polled every 10s) via
the existing `LoadingCallout` list.
- **Detail view** — clicking _View details_ on a generation swaps the
flyout body in place to render the workflow execution details for that
run.

### Key changes

- **Reusable content component**: extracted the body of
`WorkflowExecutionDetailsFlyout` into a new pure
`WorkflowExecutionDetails` component (no `EuiFlyout` chrome). The flyout
is now a thin wrapper around it, so the same content can be reused
inside the control center without nesting flyouts.
- **Delegation instead of nesting**: `LoadingCallout` gained an optional
`onViewDetails` callback. When provided, the "View details" action is
delegated upward (to the control center) rather than opening its own
flyout. Existing inline Attack Discovery behavior is unchanged when the
callback is absent.
- **Generations list threading**: `Generations` now accepts `maxItems`
(defaults to the existing 5 latest non-dismissed generations) and
forwards `onViewDetails` to each `LoadingCallout`.
- **New `GenerationsControlCenterFlyout`**: implements the master–detail
flyout, backed by a `usePollGenerations` hook and a
`getWorkflowExecutionDetailsProps` mapper that adapts an
`AttackDiscoveryGeneration` into the props expected by
`WorkflowExecutionDetails`.
- **Telemetry**: added the `Attacks Generations Control Center Opened`
event.
- **i18n**: added the strings for the new UI.

### Test coverage

- New dedicated suites for `WorkflowExecutionDetails`,
`GenerationsControlCenterFlyout`, and
`getWorkflowExecutionDetailsProps`.
- Narrowed `workflow_execution_details_flyout/index.test.tsx` to only
cover the flyout wrapper's responsibilities (chrome + prop
pass-through); the body behavior coverage moved to the new
content-component suite, so nothing is lost.
- Extended `LoadingCallout`, `Generations`, and Attacks `content` tests
to cover the new delegation, `maxItems`, button rendering, and
telemetry.

## Verification steps

1. Bootstrap and start Kibana with the security solution enabled:
   ```bash
   yarn kbn bootstrap
   yarn start
   ```
2. Navigate to **Security → Attack discovery → Attacks**.
3. Confirm a new **Generations** button appears in the header,
positioned **before** the **Run** and **Schedule** buttons.
4. Trigger one or more attack discovery generations (via **Run**) so
there is recent history.
5. Click **Generations** and verify:
   - A single flyout opens showing the list of latest generations.
- The list refreshes automatically (polling) as generation statuses
change.
6. Click **View details** on a generation and verify:
- The flyout body swaps **in place** to the workflow execution details
(no second/stacked flyout appears).
- The details match what is shown for the same generation inline on the
page.
7. Return to the list / close the flyout and confirm the flyout closes
cleanly.

## Screenshots

https://github.com/user-attachments/assets/68776014-f5f1-4ee9-b3a9-f1ec958070ad

---

_PR developed with Cursor + Claude Opus 4.8_

(cherry picked from commit 85858597ebd99cd04ba88a0d78c3a757ac7f72d9)
