+++ x-pack/solutions/security/plugins/security_solution/public/attack_discovery/pages/loading_callout/index.tsx
@@ -58,6 +58,12 @@ interface Props {
   hideActions?: boolean;
   localStorageAttackDiscoveryMaxAlerts: string | undefined;
   loadingMessage?: string;
+  /**
+   * When provided, clicking "View details" delegates to this callback instead of
+   * opening the local `WorkflowExecutionDetailsFlyout`. Used by the master-detail
+   * control center to swap its flyout body in place (avoiding a stacked flyout).
+   */
+  onViewDetails?: (executionUuid: string) => void;
   onRefresh?: () => void;
   persistedCount?: number;
   reason?: string;
@@ -91,6 +97,7 @@ const LoadingCalloutComponent: React.FC<Props> = ({
   hideActions = false,
   localStorageAttackDiscoveryMaxAlerts,
   loadingMessage,
+  onViewDetails,
   onRefresh,
   persistedCount,
   reason,
@@ -244,8 +251,16 @@ const LoadingCalloutComponent: React.FC<Props> = ({
 
   const openFlyout = useCallback(() => {
     telemetry.reportEvent(AttackDiscoveryEventTypes.ExecutionDetailsOpened, {});
+
+    // When a delegate is provided, hand off "View details" to the parent (e.g.
+    // the master-detail control center) instead of opening a nested flyout.
+    if (onViewDetails != null && executionUuid != null) {
+      onViewDetails(executionUuid);
+      return;
+    }
+
     setIsFlyoutOpen(true);
-  }, [telemetry]);
+  }, [executionUuid, onViewDetails, telemetry]);
 
   const closeFlyout = useCallback(() => {
     setIsFlyoutOpen(false);
@@ -331,7 +346,7 @@ const LoadingCalloutComponent: React.FC<Props> = ({
         </EuiFlexItem>
       </EuiFlexGroup>
 
-      {isFlyoutOpen && (
+      {onViewDetails == null && isFlyoutOpen && (
         <WorkflowExecutionDetailsFlyout
           alertsContextCount={alertsContextCount}
           approximateFutureTime={approximateFutureTime}
