+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/attacks/generations_control_center/translations.ts
@@ -1,0 +1,50 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const GENERATIONS_BUTTON = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.generationsButton',
+  {
+    defaultMessage: 'Generations',
+  }
+);
+
+export const GENERATIONS_TITLE = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.generationsTitle',
+  {
+    defaultMessage: 'Generations',
+  }
+);
+
+export const WORKFLOW_EXECUTION_DETAILS = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.workflowExecutionDetails',
+  {
+    defaultMessage: 'Workflow execution details',
+  }
+);
+
+export const BACK_TO_GENERATIONS = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.backToGenerations',
+  {
+    defaultMessage: 'Back to generations',
+  }
+);
+
+export const NO_GENERATIONS_TITLE = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.noGenerationsTitle',
+  {
+    defaultMessage: 'No recent generations',
+  }
+);
+
+export const NO_GENERATIONS_BODY = i18n.translate(
+  'xpack.securitySolution.attacks.generationsControlCenter.noGenerationsBody',
+  {
+    defaultMessage: 'Running and recently completed attack discovery generations will appear here.',
+  }
+);
